+++ spec/requests/all/in_progress/projects_controller_spec.rb
@@ -24,4 +24,14 @@ RSpec.describe All::InProgress::ProjectsController, type: :request do
       end
     end
   end
+
+  describe "#transfers_index" do
+    context "when there are no projects" do
+      it "shows a helpful message" do
+        get transfers_all_in_progress_projects_path
+
+        expect(response.body).to include "There are no in-progress transfer projects"
+      end
+    end
+  end
 end