+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - All in progress projects now includes at tab to view only conversion projects.
+- All in progress projects now includes at tab to view only transfer projects.
 
 ### Changed